+++ src/main/java/springboot/demo/mysql/vo/specs/TMFloatSpecs.java
@@ -13,7 +13,7 @@ import springboot.demo.mysql.vo.ThingModelSpecs;
 public class TMFloatSpecs extends ThingModelSpecs {
     private Integer efficientNum;
 
-    Integer min;
+    Float min;
 
-    Integer max;
+    Float max;
 }
